Basic skeleton for the home page created
The structure is in place, cosmetic changes will seal the deal. CSS entries are organized from left to right, top to bottom. Will likely use a carousel for the banner, as it's easier than a .gif. TODO: Login function, and polish information boxes underneath the banner. Start generating file structure for the backend, let's see what can be salvaged from the previous project as well.